requirements.txt
ccxt==4.2.93
pyyaml==6.0.1
python-dotenv==1.0.1
requests==2.32.3
